+++ dll/win32/advapi32_vista/DllMain.c
@@ -2,6 +2,9 @@
 #include "advapi32_vista.h"
 
 BOOL
+RegInitialize(VOID);
+
+BOOL
 WINAPI
 DllMain(HANDLE hDll,
         DWORD dwReason,
@@ -9,6 +12,14 @@ DllMain(HANDLE hDll,
 {
     /* For now, there isn't much to do */
     if (dwReason == DLL_PROCESS_ATTACH)
+    {
         DisableThreadLibraryCalls(hDll);
+
+        if (!RegInitialize())
+        {
+            return FALSE;
+        }
+    }
+
     return TRUE;
 }
